added the abort command, added a server manager area, fixed session fixed session reloading for the user, serialize team to json before inserting into redis. Rename redis module to avoid name conflict. added a linup view.
